Change type of LUT indices from size_t to uint32_t

+++ src/lut.h
@@ -7,13 +7,13 @@ class LUT
 {
 public:
     LUT() : _n_rows(0), _n_cols(0) {}
-    LUT(size_t n_rows, size_t n_cols)
+    LUT(uint32_t n_rows, uint32_t n_cols)
         : distances(n_rows * n_cols), indices(n_rows * n_cols), _n_rows(n_rows),
           _n_cols(n_cols)
     {
     }
-    LUT(size_t n_rows, size_t n_cols, const std::vector<float> &distances,
-        const std::vector<size_t> &indices)
+    LUT(uint32_t n_rows, uint32_t n_cols, const std::vector<float> &distances,
+        const std::vector<uint32_t> &indices)
         : distances(distances), indices(indices), _n_rows(n_rows),
           _n_cols(n_cols)
     {
@@ -22,24 +22,27 @@ public:
     // Eucledian distance between point i and j
     std::vector<float> distances;
     // Index of the j-th closest point from point i
-    std::vector<size_t> indices;
+    std::vector<uint32_t> indices;
 
-    float distance(size_t i, size_t j) const
+    float distance(uint32_t i, uint32_t j) const
     {
         return distances[i * _n_cols + j];
     }
-    size_t index(size_t i, size_t j) const { return indices[i * _n_cols + j]; }
-    size_t n_rows() const { return _n_rows; }
-    size_t n_cols() const { return _n_cols; }
+    uint32_t index(uint32_t i, uint32_t j) const
+    {
+        return indices[i * _n_cols + j];
+    }
+    uint32_t n_rows() const { return _n_rows; }
+    uint32_t n_cols() const { return _n_cols; }
 
-    void resize(size_t nr, size_t nc);
+    void resize(uint32_t nr, uint32_t nc);
     void print_distances() const;
     void print_indices() const;
     void normalize();
 
 protected:
-    size_t _n_rows;
-    size_t _n_cols;
+    uint32_t _n_rows;
+    uint32_t _n_cols;
 
     // Minimum weight
     const float min_weight = 1e-6f;
